remove deps
