+++ Cargo.toml
@@ -25,6 +25,7 @@ daft-functions = {path = "src/daft-functions"}
 daft-functions-json = {path = "src/daft-functions-json", default-features = false}
 daft-functions-list = {path = "src/daft-functions-list", default-features = false}
 daft-functions-temporal = {path = "src/daft-functions-temporal", default-features = false}
+daft-functions-tokenize = {path = "src/daft-functions-tokenize", default-features = false}
 daft-functions-utf8 = {path = "src/daft-functions-utf8", default-features = false}
 daft-hash = {path = "src/daft-hash", default-features = false}
 daft-image = {path = "src/daft-image", default-features = false}
@@ -171,6 +172,7 @@ members = [
   "src/daft-functions",
   "src/daft-functions-json",
   "src/daft-functions-list",
+  "src/daft-functions-tokenize",
   "src/daft-functions-utf8",
   "src/daft-hash",
   "src/daft-image",
@@ -225,6 +227,7 @@ daft-dsl = {path = "src/daft-dsl"}
 daft-functions = {path = "src/daft-functions"}
 daft-functions-json = {path = "src/daft-functions-json"}
 daft-functions-list = {path = "src/daft-functions-list"}
+daft-functions-tokenize = {path = "src/daft-functions-tokenize"}
 daft-functions-utf8 = {path = "src/daft-functions-utf8"}
 daft-hash = {path = "src/daft-hash"}
 daft-local-execution = {path = "src/daft-local-execution"}
