+++ daft/daft/__init__.pyi
@@ -1175,21 +1175,6 @@ def udf(
 def initialize_udfs(expression: PyExpr) -> PyExpr: ...
 def get_udf_names(expression: PyExpr) -> list[str]: ...
 def resolve_expr(expr: PyExpr, schema: PySchema) -> tuple[PyExpr, PyField]: ...
-def tokenize_encode(
-    expr: PyExpr,
-    tokens_path: str,
-    use_special_tokens: bool,
-    io_config: IOConfig | None,
-    pattern: str | None,
-    special_tokens: str | None,
-) -> PyExpr: ...
-def tokenize_decode(
-    expr: PyExpr,
-    tokens_path: str,
-    io_config: IOConfig | None,
-    pattern: str | None,
-    special_tokens: str | None,
-) -> PyExpr: ...
 def minhash(
     expr: PyExpr,
     num_hashes: int,
