+++ src/daft-sql/src/modules/mod.rs
@@ -1,7 +1,6 @@
 use crate::functions::SQLFunctions;
 
 pub mod aggs;
-pub mod coalesce;
 pub mod config;
 pub mod map;
 pub mod partitioning;
@@ -9,7 +8,6 @@ pub mod python;
 pub mod sketch;
 pub mod structs;
 pub mod temporal;
-pub mod utf8;
 pub mod window;
 
 pub use aggs::SQLModuleAggs;
@@ -20,7 +18,6 @@ pub use python::SQLModulePython;
 pub use sketch::SQLModuleSketch;
 pub use structs::SQLModuleStructs;
 pub use temporal::SQLModuleTemporal;
-pub use utf8::SQLModuleUtf8;
 pub use window::SQLModuleWindow;
 
 /// A [SQLModule] is a collection of SQL functions that can be registered with a [SQLFunctions] instance.
